rename interpolator_control to animation_interpolation
- this is the proper module name
