+++ xlive/Blam/Engine/animations/animation_interpolation.h
